Tests (#7)
* Initial location comparison and scoring (#1)

* Add scrapy files

* Add .pyc to .gitignore

* Largely functional scraper

* Add initial flask configuration

* .gitignore
* Database configuration
* Database schema
* Setuptools configuration
* Basic css
* Index page draft
* Basic Jinja templates

* Add scrapy spider

Scrapes relevant data from home.co.uk.
Will require some fine tuning

* Add test files

* Correct package name and version

* Implement sorting of table by headers

* Use SQLAlchemy with Postgres in flask

* Improve scraped data

* Add missing semicolon

rebase

* Include London distance and duration information

Modify sorting function to sort according to value, but to display text

* Add initial bulma files

* Update model classes to reflect new data

* Move scrapy house_scrape scraper to services dir

* Improve scraping

Detect and handle errors more cleanly
Use SQLAlchemy with PostgreSQL instead of sqlite3
Prepare to display google maps distance matrix data
Prepare to display scores
Improve logging

* Write google maps distance matrix functions

Distance and duration to London
Custom exceptions
Runner to calculate distances and durations and insert into DB
Test pages

* Service management functions

Function to calculate scores and insert into DB

* Sort table by scores by default

* Update dependencies

* Travis Configuration (#2)

* Initial Travis configuration

* Include Python 3.6

* Correct db initialisation (#3)

* Configure testing environment with pytest (#4)

* Configure database connection for tests

* Use commit instead of flush

* Add tests for postgresql configuration

* Use absolute over relative import

* Add test db travis configuration

* Add newline to end of .travis.yml

* Add test for run_google_maps_distance_matrix

* Use pytest instead of unittest with travis

* Supported on Python 3.5+

* Write client fixture

* Remove conftest comments

rebase

* Improve scrapy helper functions

* Write tests for scrapy helper functions

* Move location_list to separate file

* Write test for get_distance_matrix

* Update google maps functions

Remove unneeded functions

* Add additional google maps tests

+++ flaskr/model.py
@@ -17,7 +17,7 @@ class Location(db.Model):
 class RentalData(db.Model):
     __tablename__ = 'rental_data'
     total_properties = db.Column(db.Integer, nullable=False, default=0)
-    average_rent = db.Column(db.Integer, nullable=True)
+    average_rent = db.Column(db.Float, nullable=True)
     rent_under_250 = db.Column(db.Integer, nullable=False, default=0)
     rent_250_to_500 = db.Column(db.Integer, nullable=False, default=0)
     datetime = db.Column(db.DateTime, default=datetime.utcnow(), primary_key=True)
